Move the createOrder method to the OrderUpdateStore
